requirements.txt
aiogram==3.20.0
python-dotenv
fastapi==0.111.0
uvicorn==0.29.0